requirements.txt
PyYAML==5.3.1
schedule==0.6.0
flake8==3.8.3